+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Updated the default Python version from 3.12.3 to 3.12.4. ([#210](https://github.com/heroku/buildpacks-python/pull/210))
 - Updated setuptools from 69.5.1 to 70.0.0. ([#205](https://github.com/heroku/buildpacks-python/pull/205))
 
 ## [0.10.0] - 2024-05-03